pyproject.toml
[project]
name = "lsm-avatar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mediapipe>=0.10.21",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.1.1",
]
